+++ cmd/pomerium/main.go
@@ -14,6 +14,7 @@ import (
 	"github.com/pomerium/pomerium/authenticate"
 	"github.com/pomerium/pomerium/authorize"
 	"github.com/pomerium/pomerium/config"
+	"github.com/pomerium/pomerium/internal/frontend"
 	"github.com/pomerium/pomerium/internal/grpcutil"
 	"github.com/pomerium/pomerium/internal/httputil"
 	"github.com/pomerium/pomerium/internal/log"
@@ -136,7 +137,7 @@ func newProxyService(opt config.Options, r *mux.Router) (*proxy.Proxy, error) {
 	if err != nil {
 		return nil, err
 	}
-	r.PathPrefix("/").Handler(service.Handler)
+	r.PathPrefix("/").Handler(service)
 	return service, nil
 }
 
@@ -169,6 +170,7 @@ func newGlobalRouter(o *config.Options) *mux.Router {
 	mux.Use(middleware.Healthcheck("/ping", version.UserAgent()))
 	mux.HandleFunc("/healthz", httputil.HealthCheck)
 	mux.HandleFunc("/ping", httputil.HealthCheck)
+	mux.PathPrefix("/.pomerium/assets/").Handler(http.StripPrefix("/.pomerium/assets/", frontend.MustAssetHandler()))
 
 	return mux
 }